requirements.txt
Flask
apscheduler
requests
bs4
mysql-connector-python